Move Utf8StreamFilter to separate package
