+++ exec/kubernetes-infra/pty_handler.go
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2012-2018 Red Hat, Inc.
+// Copyright (c) 2012-2019 Red Hat, Inc.
 // This program and the accompanying materials are made
 // available under the terms of the Eclipse Public License 2.0
 // which is available at https://www.eclipse.org/legal/epl-2.0/
@@ -14,13 +14,14 @@ package kubernetes_infra
 
 import (
 	"github.com/eclipse/che-machine-exec/api/model"
+	"github.com/eclipse/che-machine-exec/output/utf8stream"
 	"k8s.io/client-go/tools/remotecommand"
 )
 
 // Kubernetes pty handler
 type PtyHandlerImpl struct {
 	machineExec *model.MachineExec
-	filter      *model.Utf8StreamFilter
+	filter      *utf8stream.Utf8StreamFilter
 }
 
 func (t PtyHandlerImpl) Read(p []byte) (int, error) {
@@ -35,7 +36,8 @@ func (t PtyHandlerImpl) Write(p []byte) (int, error) {
 	t.machineExec.Buffer.Write(filteredCharacters)
 	t.machineExec.WriteDataToWsConnections(filteredCharacters)
 
-	return len(filteredCharacters), nil
+	// Original length of the data must be returned to continue reading of the buffer correctly.
+	return len(p), nil
 }
 
 func (t PtyHandlerImpl) Next() *remotecommand.TerminalSize {
