+++ output/line-buffer/doc.go
@@ -12,10 +12,10 @@
 
 package line_buffer
 
-// This component was implemented to have ability restore exec output.
+// This component was implemented to have ability to restore exec output.
 // Clint side should have ability to reconnect to the created exec
 // and get previous output.
-// This package it's implementation ring buffer to store exec output
-// with limited amount of lines. This limit defined like max ring buffer size.
-// If exec spawn output more than ring buffer size than we rewrite previously
+// This package it's implementation of ring buffer to store exec output
+// with limited amount of lines. This limit is defined like max ring buffer size.
+// If exec spawns output more than ring buffer size than we rewrite previously
 // stored output by ring.
